+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## Unreleased
 
+### Added
+
+* JuliaHub.jl now supports HTTP.jl 2.x, in addition to 1.x. Note that HTTP.jl 2.x requires Julia 1.10 or newer. ([#171], [#173], [#PRNUM])
+
 ## Version [v0.1.18] - 2026-07-29
 
 ### Added